+++ packages/super-editor/src/components/toolbar/defaultItems.js
@@ -871,8 +871,8 @@ export const makeDefaultItems = (superToolbar, isDev = false, windowWidth, role,
 
     // Dev - tracked changes
     // toggleTrackChanges,
-    // acceptTrackedChangeBySelection,
-    // rejectTrackedChangeOnSelection,
+    acceptTrackedChangeBySelection,
+    rejectTrackedChangeOnSelection,
     // toggleTrackChangesOriginal,
     // toggleTrackChangesFinal,