+++ packages/super-editor/src/index.js
@@ -19,6 +19,8 @@ import SuperInput from './components/SuperInput.vue';
 import AIWriter from './components/toolbar/AIWriter.vue';
 import * as fieldAnnotationHelpers from './extensions/field-annotation/fieldAnnotationHelpers/index.js';
 import * as trackChangesHelpers from './extensions/track-changes/trackChangesHelpers/index.js';
+import { TrackChangesBasePluginKey } from './extensions/track-changes/plugins/index.js';
+import { CommentsPluginKey } from './extensions/comment/comments-plugin.js';
 
 const Extensions = {
   Node,
@@ -63,4 +65,7 @@ export {
 
   // External extensions classes
   Extensions,
+
+  TrackChangesBasePluginKey,
+  CommentsPluginKey,
 };