+++ packages/superdoc/src/SuperDoc.vue
@@ -122,6 +122,7 @@ const onCommentsLoaded = ({ editor, comments, replacedFile }) => {
     nextTick(() => {
       commentsStore.processLoadedDocxComments({
         superdoc: proxy.$superdoc,
+        editor,
         comments,
         documentId: editor.options.documentId
       });